+++ src/Spectrogram/SpectrogramGenerator.cs
@@ -56,7 +56,7 @@ public class SpectrogramGenerator
     /// <summary>
     /// Number of samples per second
     /// </summary>
-    public int SampleRate { get => Settings.SampleRate; }
+    public double SampleRate { get => Settings.SampleRate; }
 
     /// <summary>
     /// Number of samples to step forward after each FFT is processed.
@@ -218,7 +218,7 @@ public class SpectrogramGenerator
 
         var fftsMel = new List<double[]>();
         foreach (var fft in FFTs)
-            fftsMel.Add(FftSharp.Mel.Scale(fft, SampleRate, melBinCount));
+            fftsMel.Add(FftSharp.Mel.Scale(fft, (int)SampleRate, melBinCount));
 
         return fftsMel;
     }
